chore: move MockLimits into existing mock file
