+++ pkg/limits/mock_test.go
@@ -10,6 +10,23 @@ import (
 	"github.com/grafana/loki/v3/pkg/limits/proto"
 )
 
+type MockLimits struct {
+	MaxGlobalStreams int
+	IngestionRate    float64
+}
+
+func (m *MockLimits) MaxGlobalStreamsPerUser(_ string) int {
+	return m.MaxGlobalStreams
+}
+
+func (m *MockLimits) IngestionRateBytes(_ string) float64 {
+	return m.IngestionRate
+}
+
+func (m *MockLimits) IngestionBurstSizeBytes(_ string) int {
+	return 1000
+}
+
 type mockWAL struct {
 	t                    *testing.T
 	NumAppendsTotal      int
